HTTP/School/requirements.txt
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.4.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.21.6
openpyxl==3.1.2
pandas==1.3.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
